fix(addLiquidity): Fix addLiquidity track error
